Refactor SQL query rewriting for state history and version handling
- Introduced a new function `rewrite_lix_state_history_query` to handle the rewriting of queries involving the `lix_state_history` view, improving the handling of predicates and derived queries.
- Consolidated the query rewriting logic in `route.rs` to utilize a pass-based approach for normalization, enhancing maintainability and readability.
- Removed the `lix_state_by_version_view_read`, `lix_state_history_view_read`, and `lix_state_view_read` modules as their functionality has been integrated into the new structure.
- Updated tests to ensure correctness of the new query rewriting logic, particularly focusing on the handling of placeholders and predicate pushing.
