+++ packages/engine/src/sql/steps/mod.rs
@@ -4,11 +4,8 @@ pub mod lix_active_account_view_read;
 pub mod lix_active_account_view_write;
 pub mod lix_active_version_view_read;
 pub mod lix_active_version_view_write;
-pub mod lix_state_by_version_view_read;
 pub mod lix_state_by_version_view_write;
-pub mod lix_state_history_view_read;
 pub mod lix_state_history_view_write;
-pub mod lix_state_view_read;
 pub mod lix_state_view_write;
 pub mod lix_version_view_read;
 pub mod lix_version_view_write;
